Add an option to disable the use of the splice syscall
The splice syscall improves the performance of transfering data between
two file descriptors (in our case between two sockets) but involves the
creation of pipes. When a connection is made to the proxy, 6 file
descriptors are created (instead of just 2 when the splice syscall is
not used):
- 1 for the socket between the downstream client and the proxy
- 1 for the socket between the proxy and the upstream server
- 4 for the splice pipes
So when there are a lot of connections made to the proxy, the limit on
the number of file descriptors allocated to a process can quickly be
reached. If it is not possible to increase this limit and more
connections to the proxy are necessary, the use of the splice syscall
can now be disabled.